try one refactor option
