+++ vquest/util.py
@@ -2,7 +2,8 @@
 Misc standalone utilities.
 """
 
-from io import BytesIO
+import csv
+from io import BytesIO, StringIO
 from zipfile import ZipFile
 
 def chunker(iterator, chunksize):
@@ -32,3 +33,20 @@ def unzip(txt):
             with zipobj.open(item) as stream:
                 zipdata[item.filename] = stream.read()
     return zipdata
+
+def airr_to_fasta(
+        airr_txt,
+        seqid_col="sequence_id", aln_col="sequence_alignment", fallback_col="sequence"):
+    """Convert AIRR TSV table to FASTA, both as strings.
+
+    If the alignment column is empty for a given row, the sequence will be
+    taken from fallback_col, if provided.
+    """
+    reader = csv.DictReader(StringIO(airr_txt), delimiter="\t")
+    fasta = ""
+    for row in reader:
+        seq = row[aln_col]
+        if fallback_col:
+            seq = seq or row[fallback_col]
+        fasta += ">%s\n%s\n" % (row[seqid_col], seq)
+    return fasta
